web_app/backend/api/requirements.txt
appdirs==1.4.4
attrs==19.3.0
awscli==1.18.12
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==3.4.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.5
joblib==0.16.0
json5==0.9.4
lightgbm==2.3.1
numpy==1.18.1
pandas==1.0.3
pathlib2==2.3.5
pathspec==0.8.0
patsy==0.5.1
pickleshare==0.7.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.2
regex==2020.7.14
requests==2.22.0
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.5.2
six==1.15.0
statsmodels==0.11.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.1.0
xgboost==1.1.1
